requirements.txt
anthropic>=0.49.0
python-dotenv==1.2.2
httpx==0.28.1
pydantic==2.13.2
anyio==4.13.0
